Initial public version
This is the first public version of lib-streamlabs-ipc showcasing the performance of the raw IPC code. The performance right now is good enough for what we were aiming for, with the primary limit being the Windows Kernel, CPU speed and Memory speed. These are also the limiting factors for how many clients can connect to a single server, but usually this should be fast enough for our use case right now.

Possible improvements:
- Windows: Allow multiple servers (distributed computing).
- Windows: Use Overlapped I/O instead of synchronous I/O.
- Windows: Don't rely on process id uniqueness, as there might be more than one client in a process.
- Group messages for both sending and retrieving.
- Further reduce copy overhead of parameters.

Current bugs:
- Windows: Server gets stuck shutting down.
- Windows: Client connect triggers two pipes at the same time.