os-namedsocket: Improve stability and usability
The main class now provides most functionality that is easily implemented as a wrapper around the actual method, such as identified if it is a server or not. It is additionally now a mostly asynchronous class.

A connection object can also now outlive the parent, but the connection itself should be closed at the same time the parent Named Socket is closed.