os-namedsocket-win: Make Read/Write synchronous with timeout
The previous threading model didn't really make any sense, as the only reason for overlapped IO was to add timeouts, not add another layer of complexity. Therefore, Read() and Write() calls can now fail and will return 0 or empty buffers if the timeout is hit or another error occurs.

Care needs to be taken when sending or receiving data as it is not guaranteed that it is sent at all. I'm not sure what happens when multiple calls to Read() and Write() happen from different threads, Microsoft states that eventually GetLastError() will return ERROR_OUT_OF_MEMORY with too many overlapped IO requests.