ipc-client: Refactor and fix a freeze caused by a read operation
The previous code would incorrectly pass the buffer, which made the connection think the client wanted buffer.size() bytes. This is now fixed so that the client will always specify the exact number of bytes it wants.

Additionally some improvements were done to reduce the time that it takes to react to messages.

Bonus: The code now looks much cleaner than before