ipc-server-instance: Remove even more flush calls
Technically a finished write_event() call should be visible on the other end instantly. This is sometimes not the case and flushing would help here. However it seems that flushing can also freeze until the other end reads.

I've removed the calls to flush() completely now due to this, as a freeze is unwanted at all times.