os-signal-win: Use Semaphores instead of Events
On Windows, Events (and ConditionVariable) can get stuck, even though the threads performed the tasks asked of them. For ConditionVariable, a wait() sometimes could 'forget' about a previous signal(), thus causing a deadlock. While for Events, this is rarer, but actually a Kernel design issue that is documented and can be worked around.

Semaphores, ConditionVariables and Events basically work the same and should not have a much larger overhead over one another.