project: Add lib-datalane
lib-datalane is built to be similar to lib-streamlabs-ipc, but with a different goal. While the core of lib-streamlabs-ipc incorrectly was trying to be synchronous, lib-datalane was built from the ground up to support all modern ideas and ideologies. read(), write() and accept() are all asynchronous, everything can be waited on by the code using the kernel allowing for high performance responses, and other things. In addition to that, lib-datalane's only goal is to push as much data per second as possible, while lib-streamlabs-ipc's only goal was to get as many RPC out as possible.